Allow return false to bubble up to stop traversing
If a visitor function uses "return false" to abort traversing the ast it only works if it's currently traversing a root-element. Deeper in the tree it will not be able to break the traversing.

This fixes issue #257.

I'm honestly not sure if this change breaks anything. It works for my project which builds flawlessly. The testsuite also passes, but I don't think the test is thorough enough.