package.json compliance
As I understand things, a package.json should always have a name and a version at a minimum. In addition if the main resource is not index.js then a "main" property is required.

I don't understand Volo, and it may well be taking care of this business somehow (I see there's an npm module published), but this non-conforming package.json is confusing to an outsider.