Wiring up existing source maps of transpiled codes
When `generateSourceMaps` option is enabled, r.js now detects if the
module to be concatenated already declares a source map. It tries to
load the existing source map, and translates line numbers of all
mappings in that source map to those of the final output.

This fixes most of #470, and works well as long as `optimize` option is
set to `none`.  Some work still remains to pass a correct
`--in-source-map` option to UglifyJS when also uglifying.  Currently,
UglifyJS ignores r.js's carefully generated source map producing a bogus
one if `optimize` option is set to `uglify2`.

Used https://github.com/lydell/source-map-url (v0.2.0) for detecting
sourceMappingURL= comments from the code.