Account for wrap config in sourcemap generation
Text and files specified in a configuration as start or end
wrappers for build output are now included in sourcemap
generation.  Doing so addresses an issue in which tools
relying on the sourcemap, such as amdclean, would fail
due to invalid information in the sourcemap file.

Added somewhat-friendlier testing for sourcemap equality
in the unit tests.  This testing uses JSON.parse to load the
sourcemaps, then it compares the properties to make it
easier to determine where any unexpected results lie.