Add beforeFileRead configuration option.
This allows 'hooking' of file reads to specify content for a file
without actually doing any I/O. This achieves a similar result as the
rawText configuration option, with the added benefit that the list of
modules does not need to be known beforehand.

An exmaple use case is when optimizing a module that serves as a plugin
to a lerger framework, and where the module imports certain
functionality from the framework without these files being available at
build time.