Add excludeSpecific configuration option.
Just before the module is flattenned and built, and just after the
exclude and excludeShallow options have been applied, excludeShallow is
called once for every remaining dependency, and the dependency is
removed if the call returns true.

This allows more specific control over which modules are excluded than
exclude (module + dpendencies) and excludeShallow (module).